+++ src/core_modules/capture-core/RulesEngine/OptionSetHelper/OptionSetHelper.js
