+++ src/core_modules/capture-core/RulesEngine/VariableService/VariableService.js
