+++ src/core_modules/capture-core/RulesEngine/VariableService/variablePrefixes.const.js
