+++ src/core_modules/capture-core/RulesEngine/__tests__/runRulesForSingleEvent.test.js
@@ -1,4 +1,4 @@
-import runRulesForSingleEvent from '../../../capture-core/rulesEngineActionsCreator/runRulesForSingleEvent';
+import runRulesForSingleEvent from '../../rulesEngineActionsCreator/runRulesForSingleEvent';
 
 describe('rules engine', () => {
     const currentEvent = {};
