+++ src/core_modules/capture-core/RulesEngine/commonUtils/trimQuotes.js
