+++ src/core_modules/capture-core/RulesEngine/converters/momentConverter.js
@@ -1,6 +1,6 @@
 // @flow
 import moment from 'moment';
-import { getFormattedStringFromMomentUsingEuropeanGlyphs } from '../../date';
+import { getFormattedStringFromMomentUsingEuropeanGlyphs } from '../../../capture-core-utils/date';
 import type { IMomentConverter } from '../rulesEngine.types';
 
 const momentFormat = 'YYYY-MM-DD';
