+++ src/core_modules/capture-core/RulesEngine/converters/rulesEffectsValueConverter.js
@@ -1,7 +1,7 @@
 // @flow
 /* eslint-disable class-methods-use-this */
 import moment from 'moment';
-import { convertMomentToDateFormatString } from '../../../capture-core/utils/converters/date';
+import { convertMomentToDateFormatString } from '../../utils/converters/date';
 import type { IConvertOutputRulesEffectsValue } from '../rulesEngine.types';
 
 const dateMomentFormat = 'YYYY-MM-DD';
