+++ src/core_modules/capture-core/RulesEngine/d2Functions/d2Functions.js
