+++ src/core_modules/capture-core/RulesEngine/d2Functions/index.js
