+++ src/core_modules/capture-core/RulesEngine/rulesEffectsProcessor/processTypes.const.js
