+++ src/core_modules/capture-core/RulesEngine/rulesEffectsProcessor/rulesEffectsProcessor.js
