+++ src/core_modules/capture-core/RulesEngine/typeKeys.const.js
