+++ src/core_modules/capture-core/RulesEngine/typeToInterfaceFnName.const.js
