+++ src/core_modules/capture-core/rulesEngineActionsCreator/runRulesForTEI.js
@@ -1,7 +1,7 @@
 // @flow
 import log from 'loglevel';
 import { errorCreator } from '../../capture-core-utils';
-import { RulesEngine } from '../../capture-core-utils/RulesEngine';
+import { RulesEngine } from '../RulesEngine';
 import { TrackerProgram, DataElement, RenderFoundation } from '../metaData';
 import constantsStore from '../metaDataMemoryStores/constants/constants.store';
 import optionSetsStore from '../metaDataMemoryStores/optionSets/optionSets.store';
@@ -11,7 +11,7 @@ import type {
     TrackedEntityAttribute as TrackedEntityAttributeForRulesEngine,
     Enrollment,
     TEIValues,
-} from '../../capture-core-utils/RulesEngine/rulesEngine.types';
+} from '../RulesEngine/rulesEngine.types';
 
 const errorMessages = {
     PROGRAM_MISSING_OR_INVALID: 'Program is missing or is invalid',
